\name{pituitary}
\alias{pituitary}
\docType{data}
\title{Size of pituitary fissue}
\description{The University of Carolina conducted a study in which the size (in mm) of the pituitary fissure was measured on girls between an age of 8 and 14.
}
\usage{data(pituitary)}
\format{A data frame with 11 subjects.
}
\references{
Pothoff, R. F., & Roy, S. N. (1964). A generalized multivariate analysis of variance model useful especially for growth curve problems. Biometrika, 51, 313-326.

Robertson, T., Wright, F. T., & Dykstra, R. L. (1988). Order restricted statistical inference. New York, Wiley.
}
\examples{
data(pituitary)
}
\keyword{datasets}
