\name{posturo}
\alias{posturo}
\docType{data}
\title{Repeated posturographic measures}
\description{This dataset represents a subset from the posturographic data collected in Leitner et al. (sensory organisation test SOT).
}
\usage{data(posturo)}
\format{A data frame with 50 subjects, age as predictor and 3 repeated SOT measures as responses.
}
\references{
Leitner, C., Mair, P., Paul, B., Wick, F., Mittermaier, C., Sycha, T., & Ebenbichler, G. (2009). Reliability of posturographic measurements in the assessment of impaired sensorimotor function in chronic low back pain. Journal of Electromyography and Kinesiology, 19(3), 380-390.
}
\examples{
data(posturo)
}
\keyword{datasets}
